backend/requirements.txt
Flask ~= 2.2.3
gunicorn ~= 20.1.0
docker ~= 6.0.1
redis ~= 4.5.1
PyYAML ~= 6.0
